+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EditComponentFactory.java
@@ -22,8 +22,6 @@ import is.codion.framework.domain.entity.attribute.AttributeDefinition;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 
-import org.jspecify.annotations.Nullable;
-
 import javax.swing.JComponent;
 import java.util.Optional;
 
@@ -39,10 +37,9 @@ public interface EditComponentFactory<T, C extends JComponent> {
 	/**
 	 * Provides an input {@link ComponentValue} for editing a single attribute value for one or more entities.
 	 * @param editModel the edit model used to create foreign key input models
-	 * @param value the initial value to display
-	 * @return a new {@link ComponentValue} instance handling input for {@code attribute}
+	 * @return a new {@link ComponentValue} instance
 	 */
-	ComponentValue<T, C> component(SwingEntityEditModel editModel, @Nullable T value);
+	ComponentValue<T, C> component(SwingEntityEditModel editModel);
 
 	/**
 	 * Provides a way to override the default attribute caption, when presenting the component to the user.